+++ assets/js/views/DownloadFormView.vue
@@ -13,7 +13,7 @@ import SiteParameterInputView from './SiteParameterInputView.vue';
 import CachedCodesModel from '../CachedCodesModel.vue';
 import CodesWithKeysModel from '../CodesWithKeysModel.vue';
 import queryService from '../queryService';
-import {getQueryString} from '../utils';
+import {getQueryString, getQueryParamArray} from '../utils';
 import store from '../store/store.js';
 
 let downloadFormControllerClass = Vue.extend(DownloadFormController);
@@ -53,8 +53,7 @@ export default {
   },
   data() {
     return {
-      selectedForm: document.querySelector('#params-basic'),
-      container: this.form.querySelector('.download-box-input-div')
+      selectedForm: document.querySelector('#params-basic')
     };
   },
   methods: {
@@ -153,7 +152,8 @@ export default {
       let dataDetailsClass = Vue.extend(DataDetailsView);
       this.dataDetailsView = new dataDetailsClass({
         propsData: {
-          container: this.form.querySelector('.download-box-input-div')
+          container: this.form.querySelector('.download-box-input-div'),
+          form: this.form
           // updateResultTypeAction: (dataProfile) => {
             // Change the 'form action' on both the basic and advanced forms everytime the 'Data Profiles'
             // radio buttons are changed. This keeps the form action correctly selected when the basic form
@@ -225,18 +225,20 @@ export default {
 
       const basicForm = document.querySelector('#params-basic');
       // Set up change event handler for form inputs to update the hash part of the url
-      this.form.querySelectorAll('input[name], select[name], textarea[name], button[name]').forEach(input => {
-        input.onchange = () => {
+      // this.form.querySelectorAll('input[name], select[name], textarea[name], button[name]').forEach(input => {
+      this.form.querySelectorAll('input[name]:not([name="dataProfile"]), select[name], textarea[name], button[name]').forEach(input => {
+          input.onchange = () => {
           // console.log('in downloadview this.form ', this.form)
-          const queryParamArray = this.getQueryParamArray(this.form);
+          const queryParamArray = getQueryParamArray(this.form);
           const queryString = getQueryString(queryParamArray, ['zip', 'csrf_token']);
           window.location.hash = `#${queryString}`;
           shareText.value = window.location.href;
         };
       });
-      basicForm.querySelectorAll('input[name], select[name], textarea[name], button[name]').forEach(input => {
+      // basicForm.querySelectorAll('input[name], select[name], textarea[name], button[name]').forEach(input => {
+      basicForm.querySelectorAll('input[name]:not([name="dataProfile"]), select[name], textarea[name], button[name]').forEach(input => {
         input.onchange = () => {
-          const queryParamArray = this.getQueryParamArray(basicForm);
+          const queryParamArray = getQueryParamArray(basicForm);
           const queryString = getQueryString(queryParamArray, ['zip', 'csrf_token']);
           window.location.hash = `#${queryString}`;
           shareText.value = window.location.href;
@@ -290,6 +292,13 @@ export default {
 
       return initComplete;
     },
+    updateUrlHashString(form) {
+      const queryParamArray = getQueryParamArray(form);
+      const queryString = getQueryString(queryParamArray, ['zip', 'csrf_token']);
+      window.location.hash = `#${queryString}`;
+      shareText.value = window.location.href;
+    },
+
     updateSelected(formtype) {
       this.selectedForm = formtype;
     },
@@ -298,7 +307,7 @@ export default {
       const shareText = shareContainer.querySelector('textarea');
       let self = this;
       let getQuery = function () {
-        const queryParamArray = self.getQueryParamArray(self.selectedForm);
+        const queryParamArray = getQueryParamArray(self.selectedForm);
         const queryString = getQueryString(queryParamArray, ['zip', 'csrf_token']);
         window.location.hash = `#${queryString}`;
         shareText.value = window.location.href;
@@ -356,10 +365,11 @@ export default {
       }
       // Set up the Download button
       form.querySelector(buttonSelector).onclick = (event) => {
-        const fileFormat = this.dataDetailsView.getMimeType();
+        // const fileFormat = this.dataDetailsView.getMimeType();
         // const resultType = this.dataDetailsView.getResultType();
+        const fileFormat = store.state.mimeType;
         const dataProfile = store.state.dataProfile.mainProfile;
-        const queryParamArray = this.getQueryParamArray(form);
+        const queryParamArray = getQueryParamArray(form);
         const queryString = decodeURIComponent(getQueryString(queryParamArray));
         const self = this;
         const startDownload = (totalCount) => {
@@ -409,97 +419,96 @@ export default {
       return downloadFormController.validateDownloadForm(this.form);
     },
 
-    /*
-     * Return an array of Objects with name, value, and data-multiple attributes representing the current state
-     * of the form. Empty
-     * values are removed from the array. For selects and checkbox fieldsets that can have multiple values value will be an array, otherwise
-     * it will be a string.
-     * @return {Array of Objects with name, value, and multiple properties}
-     */
-    getQueryParamArray(currentForm) {
-      let stores = [
-        {name: 'countrycode', value: store.state.countrySelectedState},
-        {name: 'statecode', value: store.state.stateSelectedState},
-        {name: 'countycode', value: store.state.countySelectedState},
-        {name: 'siteType', value: store.state.sitetypeSelectedState},
-        {name: 'charGroup', value: store.state.chargroupSelectedState},
-        {name: 'sampleMedia', value: store.state.sampleMediaSelectedState},
-        {name: 'organization', value: store.state.orgIDSelectedState},
-        {name: 'project', value: store.state.projIDSelectedState},
-        {name: 'siteid', value: store.state.siteIDSelectedState},
-        {name: 'characteristicName', value: store.state.charSelectedState},
-        {name: 'assemblage', value: store.state.assemblageSelectedState},
-        {name: 'subjectTaxonomicName', value: store.state.taxSelectedState}
-      ];
-
-      // Need to eliminate form parameters within the mapping-div
-      const formInputs =
-          currentForm.querySelectorAll(
-              // 'input:not(#mapping-div input, #nldi-map input, input[name="dataProfile"]), textarea:not(#mapping-div textarea, #nldi-map textarea), select:not(#mapping-div select, #nldi-map select), button:not(#mapping-div button, #nldi-map button)'
-              'input:not(#mapping-div input, #nldi-map input), textarea:not(#mapping-div textarea, #nldi-map textarea), select:not(#mapping-div select, #nldi-map select), button:not(#mapping-div button, #nldi-map button)'
-
-          );
-      let queryString = [];
-      let providersArray = [];
-      formInputs.forEach(function (el, index) {
-        let multiselectArray = [];
-        if (el.type !== 'radio' || el.checked || el.className === 'datasources usa-checkbox__input') {
-
-            const value = el.value;
-            const valueIsNotEmpty = typeof value === 'string' ? value : value.length > 0;
-            const name = el.getAttribute('name');
-
-            if (valueIsNotEmpty && name && el.className != 'multiselect__input' && el.className != 'hidden-input') {
-              if (valueIsNotEmpty && el.className === 'datasources usa-checkbox__input' && el.checked === true) {
-                providersArray.push(value);
-              } else if (el.className !== 'datasources usa-checkbox__input') {
-                if(valueIsNotEmpty && name === 'dataProfile') {
-                  if(el.dataset['subprofile'] !== '') {
-                    console.log('sub profile ', el.dataset['subprofile'])
-                    queryString.push({
-                      name: name,
-                      value: el.dataset['subprofile'],
-                      multiple: el.dataset.multiple ? true : false
-                    });
-                  }
-                } else {
-                  queryString.push({
-                    name: name,
-                    value: value,
-                    multiple: el.dataset.multiple ? true : false
-                  });
-                }
-              }
-            }
-            if (index === formInputs.length - 1) {
-              queryString.push({
-                name: 'providers',
-                value: providersArray,
-                multiple: el.dataset.multiple ? true : false
-              });
-            } else if (el.className === 'hidden-input') {
-              stores.forEach(function (state) {
-                if (el.name === state.name) {
-                  state.value.forEach(function (stateValue) {
-                    multiselectArray.push(stateValue.id);
-                  });
-
-                  if (multiselectArray.length !== 0) {
-                    queryString.push({
-                      name: state.name,
-                      value: multiselectArray,
-                      multiple: el.dataset.multiple ? true : false
-                    });
-                  }
-                }
-              });
-            }
-          }
-
-      });
-      // console.log('in get query param array queryString ', queryString)
-      return queryString;
-    },
+    // /*
+    //  * Return an array of Objects with name, value, and data-multiple attributes representing the current state
+    //  * of the form. Empty
+    //  * values are removed from the array. For selects and checkbox fieldsets that can have multiple values value will be an array, otherwise
+    //  * it will be a string.
+    //  * @return {Array of Objects with name, value, and multiple properties}
+    //  */
+    // getQueryParamArray(currentForm) {
+    //   let stores = [
+    //     {name: 'countrycode', value: store.state.countrySelectedState},
+    //     {name: 'statecode', value: store.state.stateSelectedState},
+    //     {name: 'countycode', value: store.state.countySelectedState},
+    //     {name: 'siteType', value: store.state.sitetypeSelectedState},
+    //     {name: 'charGroup', value: store.state.chargroupSelectedState},
+    //     {name: 'sampleMedia', value: store.state.sampleMediaSelectedState},
+    //     {name: 'organization', value: store.state.orgIDSelectedState},
+    //     {name: 'project', value: store.state.projIDSelectedState},
+    //     {name: 'siteid', value: store.state.siteIDSelectedState},
+    //     {name: 'characteristicName', value: store.state.charSelectedState},
+    //     {name: 'assemblage', value: store.state.assemblageSelectedState},
+    //     {name: 'subjectTaxonomicName', value: store.state.taxSelectedState}
+    //   ];
+    //
+    //   // Need to eliminate form parameters within the mapping-div
+    //   const formInputs =
+    //       currentForm.querySelectorAll(
+    //           // 'input:not(#mapping-div input, #nldi-map input, input[name="dataProfile"]), textarea:not(#mapping-div textarea, #nldi-map textarea), select:not(#mapping-div select, #nldi-map select), button:not(#mapping-div button, #nldi-map button)'
+    //           'input:not(#mapping-div input, #nldi-map input), textarea:not(#mapping-div textarea, #nldi-map textarea), select:not(#mapping-div select, #nldi-map select), button:not(#mapping-div button, #nldi-map button)'
+    //
+    //       );
+    //   let queryString = [];
+    //   let providersArray = [];
+    //   formInputs.forEach(function (el, index) {
+    //     let multiselectArray = [];
+    //     if (el.type !== 'radio' || el.checked || el.className === 'datasources usa-checkbox__input') {
+    //
+    //         const value = el.value;
+    //         const valueIsNotEmpty = typeof value === 'string' ? value : value.length > 0;
+    //         const name = el.getAttribute('name');
+    //
+    //         if (valueIsNotEmpty && name && el.className != 'multiselect__input' && el.className != 'hidden-input') {
+    //           if (valueIsNotEmpty && el.className === 'datasources usa-checkbox__input' && el.checked === true) {
+    //             providersArray.push(value);
+    //           } else if (el.className !== 'datasources usa-checkbox__input') {
+    //             if(valueIsNotEmpty && name === 'dataProfile') {
+    //               if(el.dataset['subprofile'] !== '') {
+    //                 console.log('sub profile ', el.dataset['subprofile'])
+    //                 queryString.push({
+    //                   name: name,
+    //                   value: el.dataset['subprofile'],
+    //                   multiple: el.dataset.multiple ? true : false
+    //                 });
+    //               }
+    //             } else {
+    //               queryString.push({
+    //                 name: name,
+    //                 value: value,
+    //                 multiple: el.dataset.multiple ? true : false
+    //               });
+    //             }
+    //           }
+    //         }
+    //         if (index === formInputs.length - 1) {
+    //           queryString.push({
+    //             name: 'providers',
+    //             value: providersArray,
+    //             multiple: el.dataset.multiple ? true : false
+    //           });
+    //         } else if (el.className === 'hidden-input') {
+    //           stores.forEach(function (state) {
+    //             if (el.name === state.name) {
+    //               state.value.forEach(function (stateValue) {
+    //                 multiselectArray.push(stateValue.id);
+    //               });
+    //
+    //               if (multiselectArray.length !== 0) {
+    //                 queryString.push({
+    //                   name: state.name,
+    //                   value: multiselectArray,
+    //                   multiple: el.dataset.multiple ? true : false
+    //                 });
+    //               }
+    //             }
+    //           });
+    //         }
+    //       }
+    //   });
+    //   // console.log('in get query param array queryString ', queryString)
+    //   return queryString;
+    // },
 
     // getResultType() {
     //   return this.dataDetailsView.getResultType();
